Moved certain reactions language utilities into a separate class and package.
Reverted: This util package is now getting exported, instead of the
codegen.helper package. The ui project was updated to make use of that.

The getFormattedName-methods now attempt to first get the overridden
reactions segment name from the node model, and only if that fails they
attempt to get if from the overridden reactions segment directly
(possibly triggering a resolve by that).

Added javadoc.