Removed the environment ChangePropagationSpecification and ChangePropagationFileParser.
Instead, each for each segment is now a new
ChangePropagationSpecification class getting generated. Compared to
directly using the segments' executors this has the advantage that it
extends CompositeDecomposingChangePropagationSpecification (used by some
projects to register additional change processors, and splits
CompositeTransactionalChanges), and the executor can still extend
AbstractEChangePropagationSpecification (splits TransactionalChanges
into EChanges).

An abstract base class for this was added to the dslruntime project. The
naming scheme currently follows that of the executor class.
The executor class can now be package-private.

If some project contains multiple reactions files / segments it wants
its ChangePropagationSpecification to, a new segment has to be created
which simply imports those segments. The project can then use the newly
generated ChangePropagationSpecification of this composite segment.

Some projects might require the dslruntime project as additional
dependency now, in order to correctly identify that the newly generated
ChangePropagationSpecification actually implements
'ChangePropagationSpecification' (due to this interface being specified
in the abstract base class in the dslruntime project).
(The alternatives would be: Not using an abstract base class for this,
or additionally specifying the 'implements
ChangePropagationSpecification' another time in the segments' specific
ChangePropagationSpecifications)

The reactions tests were updated accordingly. The reactions segments
used in the tests were merged into a single segment. (The import
mechanism can then be tested separately)