Sub reaction generators get passed the reactions segment now and can therefore directly insert generated routines.
This is for example useful if a routine is referenced by multiple reactions or by other routines.

Also (minor): Changed the order in which the reaction segments get generated. The reactions from the participations towards the commonality are generated first now and therefore at the top of the generated reactions file. This makes it slightly easier to quickly locate these segments during development and debugging.