Participation classes can now be marked as singleton.
The marked participation class and its chain of containers will only be instantiated if they don't exist yet. The singleton and its containers act as root of the participation. The singleton is implemented by creating a correspondence with the EClass of the as singleton marked participation class.

Also: Removed the 'ifOptionalPresent' helper function from XbaseHelper since it didn't provide much simplification over creating the if expression manually whenver it is required.