Moved the ParticipationContextHelper into the language extensions.
This allows it to also be used in non-generation contexts (such as by the validator).

ParticipationRoot and ParticipationContext were moved into their own files. We use WeakHashMaps to implement the caching of participation roots and participation contexts now. ParticipationContext no longer requires a reference to the ParticipationContextHelper, since it is now a utility class with static methods.

Also: We assume now that validation ensures that every CommonalityReferenceMapping has exactly one valid referenced participation.