Fixed: Handling of participation contexts containing duplicate participation classes.
If a commonality references itself (eg. a commonality 'Package' with a reference for its sub-packages), the corresponding participation contexts may contain the same participation class both as root and as non-root class. Any code that filtered the root classes from the context's participation classes would previously also unintentionally filter the duplicate participation classes from the context's non-root classes.

This has been resolved by adding the class ContextClass, which represents a ParticipationClass together with its context specific role. During the matching of the participation context the name of the participation class with the root role gets prefixed with 'external_' in order to differentiate it and its matched object from the participation class with the non-root role.

Also fixed: Singletons are only setup for participation contexts that actually use the singleton root now.

Previously we would also setup the singletons for reference mapping participation contexts, if the referenced participation has a singleton root (even thought it is not used in that context).

Also: Various refactoring inside ParticipationRoot and ParticipationContext:
* Added ContextContainment, which wraps a regular Containment but provides access to the corresponding ContextClasses.
* Separation of root, non-root and boundary classes and containments. Those are cached.
* Moved CommonalityInsertReactionsBuilder#getManagedClasses into ParticipationContext. This is now also reused inside ParticipationMatchingReactionsBuilder.
* Added ParticipationContext#getManagedContainments.
* Added ParticipationContext#isRootContext and ParticipationContext#isForSingletonContext.
* ParticipationContext and ParticipationRoot are compared by identity now (removed generated equals and hashCode methods).
* Minor method and variable renaming.