Attribute mapping operators for multi-valued features are no longer invoked twice.
For multi-valued features we check if the new value equals the old value before we apply the change. This required evaluating the operator twice previously. We now store the operator's result value in a variable and reuse the value from there. To avoid name clashes between generated variables we wrap the code for each attribute mapping in their own block expression.