Fixed: We are not able to find domain specific classes specified by attribute operators as their commonality and participation attribute types.
These classes are located in bundles which the Commonalities language plugin is not aware of and has not declared as dependency and is therefore not able to find.
Instead we look up the corresponding EClassifiers by the instance classes' qualified names now.

Also removed the therefore no longer needed ClassUtil utility class again.