Implemented checked participation conditions.
The participation conditions which are relevant in a specific participation context are checked before we instantiate a commonality for a matched participation.

Additionally we react to attribute changes for attribute operands which are used in participation conditions: Whenever a change is detected, we check if the conditions for an already instantiated participation are still fulfilled and otherwise delete the corresponding intermediate again. And if no intermediate exists yet, we check if we can match the participation now.

Other related changes:
* Added ParticipationConditionHelper and CheckedParticipationConditionsHelper for common utilities and code related to participation conditions.
* ParticipationConditionOperatorHelper#enforce and #check explicitly return a XExpression now.
* Added a 'deleteObject' routine which simplifies object deletion from within execute actions.
* Removed a few notes about the previously no longer required matching of attribute references: The matching triggered by changes to participation condition attribute operands has to take attribute references into account.