Changes to reference mapping participation contexts.
We support declaring multiple reference mappings for the same participation / domain now. All reference mappings for the same domain are taken into account when deriving the participation root and context for the referenced participation. This change does not apply to attribute reference mappings yet.

A consequence of this is that the participation root and context may contain multiple reference mappings and multiple reference root classes now. Various code had to be adapted to this change.

The 'ContainmentTree' used by the ParticipationMatcher may principally contain multiple root nodes now. It has therefore been updated accordingly and renamed to 'ContainmentContext'. The 'ContainmentTreeBuilder' has been renamed accordingly as well.

The ParticipationMatcher doesn't actually support the matching of ContainmentContexts with multiple root nodes yet. However, the ParticipationMatcher has access to the correspondence tags of the ContainmentContext nodes now, which can be used to implement the matching in a subsequent step.
We would however first have implement the 'reference targets' ('via' keyword) for this to be useful.

We no longer support the read and write flags for reference mappings. The semantics and use cases of those are unclear currently.

Other related changes:
* The order of the ParticipationRoot#classes should no longer be relevant.
* Internal: ContainmentContext preserves the order of inserted nodes and edges now.
* Added ParticipationContext#getDeclaringReference as shortcut.
* XbaseHelper#stringLiteral(String) returns a null literal now when passed a null String.
* Added ReactionsGeneratorConventions#correspondingVariableName(ContextClass). This simplified dealing with external participation classes in CommonalityInsertReactionsBuilder.
* Added ParticipationContext#getResourceClass.
* Removed ReactionsGeneratorConventions#getReactionName(CommonalityReferenceMapping).
* Various Javadoc and code comment updates.